feat!: Restructure material 3 Instructions screen classes
