+++ pages/src/test/java/uk/gov/android/ui/pages/m3/instructions/InstructionsScreenshotM3Test.kt
@@ -1,15 +1,17 @@
-package uk.gov.android.ui.pages.instructions
+package uk.gov.android.ui.pages.m3.instructions
 
 import androidx.compose.runtime.Composable
 import com.android.resources.NightMode
 import org.junit.runner.RunWith
 import org.junit.runners.Parameterized
 import uk.gov.android.ui.pages.BaseScreenshotTest
-import uk.gov.android.ui.pages.BaseScreenshotTestM3
+import uk.gov.android.ui.pages.instructions.m3.Instructions
+import uk.gov.android.ui.pages.instructions.m3.InstructionsProvider
+import uk.gov.android.ui.pages.m3.BaseScreenshotTestM3
 
 @RunWith(Parameterized::class)
 class InstructionsScreenshotM3Test(
-    private val parameters: Pair<InstructionsM3, NightMode>
+    private val parameters: Pair<Instructions, NightMode>
 ) : BaseScreenshotTestM3(parameters.second) {
     override val generateComposeLayout: @Composable () -> Unit = {
         parameters.first.generate()
@@ -18,11 +20,11 @@ class InstructionsScreenshotM3Test(
     companion object {
         @JvmStatic
         @Parameterized.Parameters
-        fun values(): List<Pair<InstructionsM3, NightMode>> {
-            val result: MutableList<Pair<InstructionsM3, NightMode>> =
+        fun values(): List<Pair<Instructions, NightMode>> {
+            val result: MutableList<Pair<Instructions, NightMode>> =
                 mutableListOf()
 
-            InstructionsM3Provider().values.forEach(BaseScreenshotTest.applyNightMode(result))
+            InstructionsProvider().values.forEach(BaseScreenshotTest.applyNightMode(result))
 
             return result
         }
